reintroduced RX buffer to implement available/peek/read methods
Note that due to not being interrupt driven, the private method fillRxBuffer() is frequently called to see if there is new data to be added to the buffer.
It attempts to read for a few ms, to prevent underruns. Note however that this may have unwanted delays as a side effect.